Pull for beta update.

* 32-Bit Octave-4 support for MS-Windows.
* Experimental NVidia USB NVision emitter/stereo goggles support for Linux.
* AMD Volcanic Islands gpu family support for Linux and OSX.
* Auto-detection of display head assignments for Linux and OSX on AMD gpus.
* XOrgConfCreator + XOrgConfSelector assistants for Linux.
* PsychLinuxConfiguration improvements for Linux.
* Screen('DrawDots') shader based smooth point rendering support.
* Improvements to Shuffle(), contributed by GitHub user NxNiki.
* Add new kPsychGfxCapVCGood flag for ConserveVRAM settings to work around gpu precision issues.
* Add HiDPI detection code to Windows, with hopefully useful troubleshooting messages.
* PR655Init fixes for improved robustness on OSX.
* Various bug fixes and minor improvements in various places.

+++ PsychSourceGL/Source/Common/Screen/SCREENWaitBlanking.c
@@ -90,8 +90,8 @@ PsychError SCREENWaitBlanking(void)
     PsychGetCGDisplayIDFromScreenNumber(&cgDisplayID, windowRecord->screenNumber);
 
     // Get window size and vblank startline:
-    windowwidth = PsychGetWidthFromRect(windowRecord->rect);
-    windowheight = PsychGetHeightFromRect(windowRecord->rect);
+    windowwidth = (long) PsychGetWidthFromRect(windowRecord->rect);
+    windowheight = (long) PsychGetHeightFromRect(windowRecord->rect);
     vbl_startline = windowRecord->VBL_Startline;
 
     // Query duration of a monitor refresh interval: We try to use the measured interval,
